Extract revision types
