+++ src/api/revisions/types.d.ts
@@ -1,0 +1,11 @@
+export interface Revision {
+  content: string
+  timestamp: number
+  authors: string[]
+}
+
+export interface RevisionListEntry {
+  timestamp: number
+  length: number
+  authors: string[]
+}
